don't touch source tree if `makepkg --noextract' is invoked

+++ PKGBUILD
@@ -27,14 +27,17 @@ build(){
     _goroot=$srcdir/$_hgrepo
     _gobin=$_goroot/bin
     _goos=linux
-    if [ -d $_goroot/.hg ]; then
-        # Remove all files not under version control
-        hg --cwd $_goroot status -i -u | cut -d' ' -f2 | xargs rm -rf
-        hg --cwd $_goroot pull
-        hg --cwd $_goroot update -C -r $pkgver || return 1
-    else
-        hg clone $_hgroot $_goroot || return 1
-        hg --cwd $_goroot update -C -r $pkgver || return 1
+    # Keep source tree as is, if --noextract is given
+    if [ "$NOEXTRACT" -ne 1 ]; then
+        if [ -d $_goroot/.hg ]; then
+            # Remove all files not under version control
+            hg --cwd $_goroot status -i -u | cut -d' ' -f2 | xargs rm -rf
+            hg --cwd $_goroot pull
+            hg --cwd $_goroot update -C -r $pkgver || return 1
+        else
+            hg clone $_hgroot $_goroot || return 1
+            hg --cwd $_goroot update -C -r $pkgver || return 1
+        fi
     fi
     mkdir -p $_gobin
     cd $_goroot/src
